requirements.txt
admin==0.0.1
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clarifai==2.6.2
click==8.0.1
colorama==0.4.4
configparser==3.8.1
crayons==0.4.0
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.2.7
django-crispy-forms==1.13.0
django-import-export==2.6.1
django-multiupload==0.6.1
django-summernote==0.8.11.6
docopt==0.6.2
EasyProcess==0.3
emoji==1.2.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.36.1
html5==0.0.9
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==3.4.0
importlib-resources==5.1.0
iniconfig==1.1.1
instabot==0.117.0
instapy==0.6.13
ipp==2021.2.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lib==3.0.0
linkedin-api==1.1.0
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.3.4
MeaningCloud-python==2.0.0
mistune==0.8.4
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
num2words==0.5.10
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.9
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pdfminer.six==20200517
pdfplumber==0.5.28
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
plyer==2.0.0
proglog==0.1.9
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.7
py==1.10.0
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
PyGetWindow==0.0.9
Pygments==2.8.1
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-pptx==0.6.19
python-telegram-bot==13.4.1
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywinpty==1.0.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.12.1
schedule==1.0.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
tabula==1.0.5
terminado==0.9.4
testpath==0.4.4
toml==0.10.2
tornado==6.1
tqdm==4.57.0
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.2
Wand==0.6.6
wcwidth==0.2.5
webdriver-manager==3.2.2
webdriverdownloader==1.1.0.3
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1
XlsxWriter==3.0.1
xlwt==1.3.0
zipp==3.4.0
